requirements.txt
python-consul
bottle
gripcontrol
pyzmq
tnetstring
